requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-4-jet==1.0.9
django-autoslug==1.9.9
django-ckeditor-5==0.2.10
django-import-export==3.3.1
django-quill-editor==0.1.40
django-rangefilter==0.9.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
MarkupPy==1.14
multidict==6.0.5
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
twilio==8.12.0
typing_extensions==4.8.0
urllib3==2.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
